feat: update borrow history export
