+++ apps/frontend/src/app/3_organisms/BorrowHistory/components/CloseWithSwapLoanFrame/hooks/useGetCloseWithSwapLoans.ts
@@ -85,10 +85,9 @@ export const useGetCloseWithSwapLoans = (
   const exportData = useCallback(async () => {
     const { data } = await getCloseWithSwaps({
       variables: {
+        ...config,
         skip: 0,
         pageSize: EXPORT_RECORD_LIMIT,
-        loanIds,
-        orderBy: CloseWithSwap_OrderBy.Timestamp,
       },
     });
     let list = data?.closeWithSwaps || [];
@@ -119,7 +118,7 @@ export const useGetCloseWithSwapLoans = (
         TXID: tx.transaction.id,
       };
     });
-  }, [addNotification, getCloseWithSwaps, loanIds]);
+  }, [addNotification, config, getCloseWithSwaps]);
 
   return {
     loading: loading || loadingLoanIds,
