+++ apps/frontend/src/app/3_organisms/BorrowHistory/components/CollateralSurplusFrame/hooks/useGetCollateralSurplusWithdrawals.ts
@@ -1,13 +1,26 @@
-import { useMemo } from 'react';
+import { useCallback, useMemo } from 'react';
 
-import { OrderOptions } from '@sovryn/ui';
+import { t } from 'i18next';
+import { nanoid } from 'nanoid';
 
+import { NotificationType, OrderOptions } from '@sovryn/ui';
+
+import {
+  BITCOIN,
+  BTC_RENDER_PRECISION,
+} from '../../../../../../constants/currencies';
+import { EXPORT_RECORD_LIMIT } from '../../../../../../constants/general';
+import { useNotificationContext } from '../../../../../../contexts/NotificationContext';
+import { translations } from '../../../../../../locales/i18n';
 import { zeroClient } from '../../../../../../utils/clients';
 import {
   CollSurplusChange_Filter,
   CollSurplusChange_OrderBy,
+  useGetCollSurplusChangesLazyQuery,
   useGetCollSurplusChangesQuery,
 } from '../../../../../../utils/graphql/zero/generated';
+import { dateFormat } from '../../../../../../utils/helpers';
+import { formatValue } from '../../../../../../utils/math';
 
 export const useGetCollateralSurplusWithdrawals = (
   account: string,
@@ -15,6 +28,8 @@ export const useGetCollateralSurplusWithdrawals = (
   page: number,
   orderOptions: OrderOptions,
 ) => {
+  const { addNotification } = useNotificationContext();
+
   const config = useMemo(
     () => ({
       skip: page * pageSize,
@@ -49,5 +64,43 @@ export const useGetCollateralSurplusWithdrawals = (
     }));
   }, [data]);
 
-  return { loading, data: collSurplusChanges, refetch };
+  const [getCollSurplusChanges] = useGetCollSurplusChangesLazyQuery({
+    client: zeroClient,
+  });
+
+  const exportData = useCallback(async () => {
+    const { data } = await getCollSurplusChanges({
+      variables: {
+        ...config,
+        skip: 0,
+        pageSize: EXPORT_RECORD_LIMIT,
+      },
+    });
+    let list = data?.collSurplusChanges || [];
+
+    if (!list || !list.length) {
+      addNotification({
+        type: NotificationType.warning,
+        title: t(translations.common.tables.actions.noDataToExport),
+        content: '',
+        dismissible: true,
+        id: nanoid(),
+      });
+    }
+
+    return list.map(tx => ({
+      timestamp: dateFormat(tx.transaction.timestamp),
+      collateralChange: formatValue(
+        Math.abs(Number(tx.collSurplusChange)),
+        BTC_RENDER_PRECISION,
+      ),
+      collateralChangeToken: BITCOIN,
+      transactionType: t(
+        translations.collateralSurplusHistory.table.withdrawSurplus,
+      ),
+      transactionID: tx.transaction.id,
+    }));
+  }, [addNotification, config, getCollSurplusChanges]);
+
+  return { loading, data: collSurplusChanges, refetch, exportData };
 };
