+++ apps/frontend/src/app/3_organisms/BorrowHistory/components/LiquidationLoanFrame/hooks/useGetLiquidationLoans.ts
@@ -85,10 +85,9 @@ export const useGetLiquidationLoans = (
   const exportData = useCallback(async () => {
     const { data } = await getLiquidates({
       variables: {
+        ...config,
         skip: 0,
         pageSize: EXPORT_RECORD_LIMIT,
-        loanIds,
-        orderBy: Liquidate_OrderBy.Timestamp,
       },
     });
     let list = data?.liquidates || [];
@@ -115,7 +114,7 @@ export const useGetLiquidationLoans = (
       loanId: tx.loanId.id,
       TXID: tx.transaction.id,
     }));
-  }, [addNotification, getLiquidates, loanIds]);
+  }, [addNotification, config, getLiquidates]);
 
   return {
     loading: loading || loadingLoanIds,
