+++ apps/frontend/src/app/3_organisms/BorrowHistory/components/RolloverLoanFrame/hooks/useGetRolloverLoans.ts
@@ -85,10 +85,9 @@ export const useGetRolloverLoans = (
   const exportData = useCallback(async () => {
     const { data } = await getRollovers({
       variables: {
+        ...config,
         skip: 0,
         pageSize: EXPORT_RECORD_LIMIT,
-        loanIds,
-        orderBy: Rollover_OrderBy.Timestamp,
       },
     });
     let list = data?.rollovers || [];
@@ -115,7 +114,7 @@ export const useGetRolloverLoans = (
       loanId: tx.loanId.id,
       TXID: tx.transaction.id,
     }));
-  }, [addNotification, getRollovers, loanIds]);
+  }, [addNotification, config, getRollovers]);
 
   return {
     loading: loading || loadingLoanIds,
